+++ src/lib.rs
@@ -14,13 +14,14 @@ pub use crate::{
     browser::dom::cast::{
         to_html_el, to_input, to_kbevent, to_mouse_event, to_select, to_textarea,
     },
-    browser::service::fetch::{Method, Request},
+    browser::service::fetch::{Method, Request, ResponseDataResult, ResponseResult},
     browser::service::routing::push_route,
     browser::url::Url,
     browser::util::{
         self, body, canvas, canvas_context_2d, cookies, document, error, history, html_document,
         log, update, window,
     },
+    virtual_dom::{attrs::Attrs, listener::Listener, style::Style},
 };
 use wasm_bindgen::{closure::Closure, JsCast};