+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 
 - [BREAKING] `MessageMapper::map_message` changed to `MessageMapper::map_msg`.
 - [BREAKING] `fetch` and `storage` moved to `seed::browser::service::{fetch, storage}`.
+- Added support for `Vec<Attr>` and `Vec<Style>` in view macros.
 - `App` included in `prelude`.
 - [BREAKING] Seed refactored to use `async/.await`. `fetch.rs` docs updated.
+- Export `Attrs`, `Style`, `Listener`. ie, can import with `seed::Style` etc.
 
 ## v0.5.0
 - Added helper `seed::canvas()`, and `seed::canvas_context()` helper functions.